FloorTiles can now be 'killed'
Until we know exactly how we want the map to act I have put in just a
simple falling mechanic to show how we can simply call a 'kill' function
to manipulate the state of a specific tile.

Optional extensions could be making the tile 'sink' while the player is
stood on it until they hop off, or alternatively, make it crack from the
pressure (no displacing required)

For now, floor tiles are seperate from the tiles above them as in theory
there would be no need for the non traversable tiles to fall. They can
still be 'killed', but the objects they carry must be considered so that
they don't float.

Eventually, a tile will have a general state to determine the player's
presence. For now though not much else can be done.