Tiles fall on player move!!
LIT PATCH NOTES :100: :100:

- When the player steps on the tile, it sinks slowly (CPL: customisable
per level)
- If the player stays on for too long(CPL), they sink even faster (CPL)
and cannot escape
- After staying on the faster sinking tile for some time, they die and
the level respawns (and is differentiable from a win)
- If the player escapes in time, the tile then falls as if the player
was still on it
- If the player then goes back onto a fallen tile, they fall at the
accelerated speed to their demise
- TileFloors now have an optional 'protection' value. This allows the
player to tread on the same tile n times. This is tested and works, but
is not in the current state of the game.
- A protection value of -1 makes it never expire!
- The player can still die from staying on a protected tile for too
long!

- (Theory) blocks use the same functions as the player to sit, so they
should fall too!

~ Fixed the 'jumping' the player does on an invalid move - it returned
the previous pos of the player, but with everything floating this is due
to change every frame, so instead fetches the position it SHOULD be
rather than WAS! Clever Eh? :+1: